+++ config/karma.conf.js
@@ -24,8 +24,8 @@ module.exports = function(config) {
         // preprocess matching files before serving them to the browser
         // available preprocessors: https://npmjs.org/browse/keyword/karma-preprocessor
         preprocessors: {
-            'test/**/*.common.js': ['webpack', 'coverage'],
-            'test/**/*.browser.js': ['webpack', 'coverage']
+            'test/**/*.common.js': ['webpack', 'coverage', 'sourcemap'],
+            'test/**/*.browser.js': ['webpack', 'coverage', 'sourcemap']
         },
         reporters: ['progress', 'coverage'],
         coverageReporter: {
@@ -52,7 +52,8 @@ module.exports = function(config) {
             'karma-chrome-launcher',
             'karma-firefox-launcher',
             'karma-ie-launcher',
-            'karma-coverage'
+            'karma-coverage',
+			'karma-sourcemap-loader'
         ],
         // Start these browsers, currently available:
         // - Chrome